requirements.txt
asgiref==3.4.1
astroid==2.8.5
asttokens==2.0.5
Babel==2.9.1
black==21.11b0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.45.0
django-cleanup==5.2.0
django-cors-headers==3.10.1
django-countries==7.2.1
django-crispy-forms==1.13.0
django-crontab==0.7.1
django-doctor==2.6.0
django-environ==0.8.1
django-hosts==4.0
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
django-timezone-field==4.2.1
django-tinymce==3.4.0
djangorestframework==3.12.4
dparse==0.5.1
idna==3.2
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
numpy==1.22.0
oauthlib==3.1.1
opencv-python
packaging==21.2
pathspec==0.9.0
phonenumbers==8.12.33
Pillow==9.0.1
platformdirs==2.4.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.4.0
pylint==2.11.1
pyparsing==2.4.7
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.14.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
wrapt==1.13.3
